Abstract away inquirer boilerplate.
There's a couple things that `init` does special to deal with inquirer. Abstract that away into our own prompt function .
